+++ src/main/java/com/example/yanolja/domain/user/dto/EmailRequest.java
@@ -1,4 +1,4 @@
-package com.example.yanolja.domain.user.request;
+package com.example.yanolja.domain.user.dto;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
